+++ BTCeClient/src/main/java/com/QuarkLabs/BTCeClient/api/Api.java
@@ -34,7 +34,6 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
 
@@ -336,7 +335,7 @@ public class Api {
      */
     @NonNull
     @WorkerThread
-    public CallResult<CancelOrderResponse> cancelOrder(int orderId) {
+    public CallResult<CancelOrderResponse> cancelOrder(long orderId) {
 
         try {
             Map<String, String> parameters = new HashMap<>();
